+++ src/Card/CardFooterItem/CardFooterItem.js
@@ -6,15 +6,9 @@ export default function CardFooterItem({
   children,
   className,
   component,
-  link,
   ...props
 }) {
-  if (process.env.NODE_ENV === 'development' && link) {
-    // eslint-disable-next-line
-    console.warn('Using the `link` prop is deprecated. Use `component="a"` instead.');
-  }
-
-  const Element = link ? 'a' : component;
+  const Element = component;
 
   const classes = classNames('card-footer-item', className);
 
@@ -29,15 +23,10 @@ CardFooterItem.propTypes = {
   className: PropTypes.string,
   children: PropTypes.node,
   component: PropTypes.elementType,
-  /**
-    Deprecation Warning: Use `component="a"`.
-  */
-  link: PropTypes.bool,
 };
 
 CardFooterItem.defaultProps = {
   className: undefined,
   children: null,
   component: 'div',
-  link: false,
 };