+++ src/Dropdown/DropdownItem/DropdownItem.js
@@ -6,15 +6,9 @@ export default function DropdownItem({
   children,
   className,
   component,
-  link,
   ...props
 }) {
-  if (process.env.NODE_ENV === 'development' && link) {
-    // eslint-disable-next-line
-    console.warn('Using the `link` prop is deprecated. Use `component="a"` instead.');
-  }
-
-  const Element = link ? 'a' : component;
+  const Element = component;
 
   const classes = classNames('dropdown-item', className);
 
@@ -29,15 +23,10 @@ DropdownItem.propTypes = {
   className: PropTypes.string,
   children: PropTypes.node,
   component: PropTypes.elementType,
-  /**
-    Deprecation Warning: Use `component="a"`.
-  */
-  link: PropTypes.bool,
 };
 
 DropdownItem.defaultProps = {
   className: undefined,
   children: null,
   component: 'div',
-  link: false,
 };