+++ src/Level/LevelItem/LevelItem.js
@@ -6,15 +6,9 @@ export default function LevelItem({
   children,
   className,
   component,
-  link,
   ...props
 }) {
-  if (process.env.NODE_ENV === 'development' && link) {
-    // eslint-disable-next-line
-    console.warn('Using the `link` prop is deprecated. Use `component="a"` instead.');
-  }
-
-  const Element = link ? 'a' : component;
+  const Element = component;
   const classes = classNames('level-item', className);
 
   return (
@@ -28,15 +22,10 @@ LevelItem.propTypes = {
   className: PropTypes.string,
   children: PropTypes.node,
   component: PropTypes.elementType,
-  /**
-    Deprecation Warning: Use `component="a"`.
-  */
-  link: PropTypes.bool,
 };
 
 LevelItem.defaultProps = {
   className: undefined,
   children: null,
   component: 'div',
-  link: false,
 };