+++ src/Tag/Tag.js
@@ -8,17 +8,11 @@ export default function Tag({
   color,
   component,
   light,
-  link,
   rounded,
   size,
   ...props
 }) {
-  if (process.env.NODE_ENV === 'development' && link) {
-    // eslint-disable-next-line
-    console.warn('Using the `link` prop is deprecated. Use `component="a"` instead.');
-  }
-
-  const Element = link ? 'a' : component;
+  const Element = component;
 
   // Colors
   const colorMap = {
@@ -63,10 +57,6 @@ Tag.propTypes = {
   className: PropTypes.string,
   children: PropTypes.node,
   component: PropTypes.elementType,
-  /**
-    Deprecation Warning: Use `component="a"`.
-  */
-  link: PropTypes.bool,
   color: PropTypes.oneOf([
     'primary',
     'link',
@@ -89,7 +79,6 @@ Tag.defaultProps = {
   className: undefined,
   children: null,
   component: 'span',
-  link: false,
   color: undefined,
   light: false,
   size: undefined,