+++ src/Title/Title.js
@@ -9,33 +9,11 @@ export default function Title({
   size,
   spaced,
   subtitle,
-  header,
   ...props
 }) {
-  if (process.env.NODE_ENV === 'development' && header) {
-    // eslint-disable-next-line
-    console.warn('Using the `header` prop is deprecated. Use the `component` prop instead.');
-  }
-
   const isTitle = !subtitle ? 'title' : 'subtitle';
 
-  let Element = component;
-
-  if (header) {
-    if (size === '1') {
-      Element = 'h1';
-    } else if (size === '2') {
-      Element = 'h2';
-    } else if ((size === '3') || (!subtitle && !size)) {
-      Element = 'h3';
-    } else if (size === '4') {
-      Element = 'h4';
-    } else if ((size === '5') || (subtitle && !size)) {
-      Element = 'h5';
-    } else if (size === '6') {
-      Element = 'h6';
-    }
-  }
+  const Element = component;
 
   const sizeMap = {
     1: 'is-1',
@@ -65,10 +43,6 @@ Title.propTypes = {
   subtitle: PropTypes.bool,
   size: PropTypes.oneOf(['1', '2', '3', '4', '5', '6']),
   spaced: PropTypes.bool,
-  /**
-    Deprecation Warning: Use `component`.
-  */
-  header: PropTypes.bool,
 };
 
 Title.defaultProps = {
@@ -78,5 +52,4 @@ Title.defaultProps = {
   subtitle: false,
   size: undefined,
   spaced: false,
-  header: false,
 };